+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Flight Tracker React
 
 ## v.1.0.0
-* *Feature*: adding animation to polyline.  Adding svg icons to images
+* *Feature*: adding rotation and positional updates to plane. Updating currentMapCenter to follow plane.
+* *Feature*: adding animation to polyline. Adding custom svg icons from images.
 * *Feature*: addWaypoints method
 * *Feature*: adding GreatCircleArc poc
 * *Fix*: removing notes folder